Completed popup for daily progress

+++ iTrainee.Models/DailyProgress.cs
@@ -22,6 +22,6 @@ namespace iTrainee.Models
 
         public string MentorComments { get; set; }
 
-        public float Percentage { get; set; }
+        public int Percentage { get; set; }
     }
 }
